release

+++ internal/models/payment.go
@@ -2,14 +2,14 @@ package models
 
 // Структура оплаты
 type Payment struct {
-	Transaction string `json:"transaction"`
-	RequestID   string `json:"request_id"`
-	Currency    string `json:"currency"`
-	Provider    string `json:"provider"`
-	Amount      int    `json:"amount"`
-	PaymentDt   int    `json:"payment_dt"`
-	Bank        string `json:"bank"`
+	Transaction  string `json:"transaction"`
+	RequestID    string `json:"request_id"`
+	Currency     string `json:"currency"`
+	Provider     string `json:"provider"`
+	Amount       int    `json:"amount"`
+	PaymentDt    int64  `json:"payment_dt"`
+	Bank         string `json:"bank"`
 	DeliveryCost int    `json:"delivery_cost"`
-	GoodsTotal  int    `json:"goods_total"`
-	CustomFee   int    `json:"custom_fee"`
+	GoodsTotal   int    `json:"goods_total"`
+	CustomFee    int    `json:"custom_fee"`
 }
